requirements.txt
langchain-google-genai
langchain
streamlit
upstash-vector
google-cloud-aiplatform
exa_py
pinecone-client